Fix production environment issues.
The production build of Ember does not provide the Handlebars compiler
which several of the new libraries were depending on. We're not actually
using anything from Bootstrap for Ember so Bootstrap is now included
manually via Bower. Ember-breadcrumbs has been included into the source
tree and modified to remove the call to `Ember.Handlebars.compile` since
we're not using the default template anyway.